cloud-functions/vision-service/requirements.txt
google-cloud-vision==3.4.4
functions-framework==3.4.0
flask==2.3.3
